requirements.txt
absl-py==2.1.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.1.4
astropy-iers-data==0.2024.10.14.0.32.55
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==22.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==1.2.0
fastjsonschema==2.16.3
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
gast==0.4.0
git-filter-repo==2.45.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
idna==3.4
ipykernel==6.22.0
ipympl==0.9.3
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-widgets==3.0.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
mysql-connector-python==8.0.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
overrides==7.7.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.14.0
prometheus_client==0.21.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyerfa==2.0.1.4
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==25.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
stack-data==0.6.2
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.2
tornado==6.2
traitlets==5.9.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.7
wrapt==1.16.0
